Refactor and move to new package.
