+++ pkg/kudoctl/cmd/install.go
@@ -7,6 +7,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/kudobuilder/kudo/pkg/kudoctl/cmd/install"
+	"github.com/kudobuilder/kudo/pkg/kudoctl/cmd/params"
 )
 
 var (
@@ -43,7 +44,7 @@ func newInstallCmd(fs afero.Fs) *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// Prior to command execution we parse and validate passed arguments
 			var err error
-			options.Parameters, err = install.GetParameterMap(fs, parameters, parameterFiles)
+			options.Parameters, err = params.GetParameterMap(fs, parameters, parameterFiles)
 			if err != nil {
 				return fmt.Errorf("could not parse parameters: %v", err)
 			}
