+++ pkg/kudoctl/cmd/params/parse_test.go
@@ -1,4 +1,4 @@
-package install
+package params
 
 import (
 	"os"
