+++ pkg/kudoctl/cmd/upgrade.go
@@ -8,6 +8,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/kudobuilder/kudo/pkg/kudoctl/cmd/install"
+	"github.com/kudobuilder/kudo/pkg/kudoctl/cmd/params"
 	"github.com/kudobuilder/kudo/pkg/kudoctl/env"
 	pkgresolver "github.com/kudobuilder/kudo/pkg/kudoctl/packages/resolver"
 	"github.com/kudobuilder/kudo/pkg/kudoctl/util/kudo"
@@ -52,7 +53,7 @@ func newUpgradeCmd(fs afero.Fs) *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// Prior to command execution we parse and validate passed arguments
 			var err error
-			options.Parameters, err = install.GetParameterMap(fs, parameters, parameterFiles)
+			options.Parameters, err = params.GetParameterMap(fs, parameters, parameterFiles)
 			if err != nil {
 				return fmt.Errorf("could not parse parameters: %v", err)
 			}
